# Finalizer optimization: Replace hashmap