Problems we had by using hashmap:
 1. Hashmap is not thread safe and we access the same hashmap in
 finalizer thread and caller thread of Row's constructor.
 2. Removing reference from the hashmap is a little bit slower when
 there are many references waited to be GCed.

Refactor the NativeObjectReference, make it as a final class and calling
native resource's deallocation directly inside the class. This is needed
for further optimization, like batch delete native pointers to avoid
multiple JNI calls.

* On Genymotion Nexus 6 5.1.0 this improves around 50%
* On Huawei Honor 7 Android M this only improves **-2**% - 9%
* On ARMv7 4.1.2 this improves around 30%

Some benchmark comparison using the array solution vs hashmap solution:
(Benchmark running on Genymotion Nexus 6 5.1.0)

```
(hashmap) Creating 200000 rows - in avg 10000 rows takes 65ms
(array) Creating 200000 rows - in avg 10000 rows takes 67ms
(hashmap) Deleting 200000 rows - in avg 10000 rows takes 8ms
(array) Deleting 200000 rows - in avg 10000 rows takes 4ms

(hashmap) Creating 2000000 rows - in avg 10000 rows takes 30ms
(array) Creating 2000000 rows - in avg 10000 rows takes 28ms
(hashmap) Deleting 200000 rows - in avg 10000 rows takes 95ms
(array) Deleting 200000 rows - in avg 10000 rows takes 46ms

(hashmap) Creating 10000000 rows - in avg 10000 rows takes 33ms
(array) Creating 10000000 rows - in avg 10000 rows takes 28ms
(hashmap) Deleting 1000000 rows - in avg 10000 rows takes 62ms
(array) Deleting 1000000 rows - in avg 10000 rows takes 23ms
```