@beeender 2016-03-29
Partial fix of #2496 @realm/java