@stk1m1 2016-03-29
I really love the way native references gets handled in `ReferencesPool`, when you invoke `addReference()` and `executeDelayedDisposal()`. Nice work! :+1: