@cmelchior 2016-03-31
Apart from a small comment, this looks :+1: 
Should we add something to the changelog? Even though it is not really a user facing change, it was still triggered by a user bug?