@beeender 2016-04-01
@realm/java change log & comments updated